==> Lessons/09_DataVisualization.Rmd <==
---
title: "9: Data Visualization"
author: "Environmental Data Analytics | Kateri Salk"
date: "Spring 2019"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## LESSON OBJECTIVES
1. Perform simple data visualizations in the R package `ggplot`
2. Develop skills to adjust aesthetics and layers in graphs
3. Apply a decision tree framework for appropriate graphing methods

## SET UP YOUR DATA ANALYSIS SESSION

```{r, message = FALSE, warning = FALSE}
getwd()
library(tidyverse)

PeterPaul.chem.nutrients <- read.csv("./Data/Processed/NTL-LTER_Lake_Chemistry_Nutrients_PeterPaul_Processed.csv")
PeterPaul.nutrients.gathered <- read.csv("./Data/Processed/NTL-LTER_Lake_Nutrients_PeterPaulGathered_Processed.csv")
PeterPaul.chem.nutrients.summaries <- read.csv("./Data/Processed/NTL-LTER_Lake_Summaries_PeterPaul_Processed.csv")
EPAair <- read.csv("./Data/Processed/EPAair_O3_PM25_NC1718_Processed.csv.csv")

EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")
PeterPaul.chem.nutrients$sampledate <- as.Date(PeterPaul.chem.nutrients$sampledate, format = "%Y-%m-%d")

```

## GGPLOT

ggplot, called from the package `ggplot2`, is a graphing and image generation tool in R. This package is part of tidyverse. While base R has graphing capabilities, ggplot has the capacity for a wider range and more sophisticated options for graphing. 
ggplot has only a few rules: 

* The first line of ggplot code always starts with `ggplot()`
* A data frame must be specified within the `ggplot()` function. Additional datasets can be specified in subsequent layers.
* Aesthetics must be specified, most commonly x and y variables but including others. Aesthetics can be specified in the `ggplot()` function or in subsequent layers.
* Additional layers must be specified to fill the plot. 

### Geoms

Here are some commonly used layers for plotting in ggplot: 

* geom_bar
* geom_histogram
* geom_freqpoly
* geom_boxplot
* geom_violin
* geom_dotplot
* geom_point
* geom_errorbar
* geom_smooth
* geom_line
* geom_area
* geom_abline (plus geom_hline and geom_vline)
* geom_text

### Aesthetics
Here are some commonly used aesthetic types that can be manipulated in ggplot: 

* color
* fill
* shape
* size
* transparency

### Plotting continuous variables over time: Scatterplot
```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
ggplot(EPAair, aes(x = Date, y = Ozone)) +  #inside aesthetic put x,y
  geom_point() #called a layer, within this layer, specify asethetics;
#geom_point(df2,), you can have another dataframe with some factor matches.
O3plot <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone))
print(O3plot)

# Fix this code # 在geom后面aes后面的括号里，都是base on dataset的，假如color blue是一个general的东西，就给它放在括号外面就好了。
O3plot2 <- ggplot(EPAair) +
  geom_point(aes(x = Date, y = Ozone, color = 'blue'))+
scale_color_manual(values = c('blue' = 'blue'))
print(O3plot2)

# Add additional variables
library(lubridate)
EPAair<-
  EPAair%>%
 mutate(month = month(Date)) %>%
  mutate(year=year(Date))
PMplot <- 
  ggplot(EPAair, aes(x = month, y = PM2.5, shape = as.factor(year), color = Site.Name)) +
  geom_point()
print(PMplot)

# Separate plot with facets #One column with different levels
PMplot.faceted <-
  ggplot(EPAair, aes(x = month, y = PM2.5, shape = as.factor(year))) +
  geom_point() +
  facet_wrap(vars(Site.Name), nrow = 3)
print(PMplot.faceted)

PMplot.faceted2 <-
  ggplot(EPAair, aes(x = month, y = PM2.5)) +
  geom_point() +
  facet_grid(Site.Name ~ year)
print(PMplot.faceted2)

# Filter dataset within plot building #take subset in the first line. The first line after GGplot should always indicate a dataset
O3plot.Blackstone <- 
  ggplot(subset(EPAair, Site.Name == "Blackstone"), aes(x = Date, y = Ozone)) + 
  geom_point() +
  geom_line()
print(O3plot.Blackstone)

# Exercise: build your own scatterplots of PeterPaul.chem.nutrients

# 1. 
# Plot surface temperatures by day of  year. 
# Color your points by year, and facet by lake in two rows.
PeterTemp <- 
  ggplot(subset(PeterPaul.chem.nutrients,depth=0),
aes(x = daynum, y = temperature_C, color= year4))+
  geom_point()+
  facet_wrap(vars(lakename),nrow = 2)
print(PeterTemp)

#2. 
# Plot temperature by date. Color your points by depth.
# Change the size of your point to 0.5
ggplot(PeterPaul.chem.nutrients, aes(x=
                                       sampledate, y=temperature_C,
                                     color=depth))+
  geom_point(size=0.5)

```
### Plotting the relationship between two continuous variables: Scatterplot

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
# Scatterplot
lightvsDO <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point()
print(lightvsDO)

# Adjust axes #lots of light, phtotosynthesis, exchange with the outside oxygen
lightvsDOfixed <- 
  ggplot(PeterPaul.chem.nutrients, aes(x = irradianceWater, y = dissolvedOxygen)) +
  geom_point() +
  xlim(0, 250) +
  ylim(0, 20)
print(lightvsDOfixed)

# Depth in the fields of limnology and oceanography is on a reverse scale
tempvsdepth <- 
  #ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth)) +
  ggplot(PeterPaul.chem.nutrients, aes(x = temperature_C, y = depth, color = daynum)) +
  geom_point() +
  scale_y_reverse() #make a reverse on y axes
print(tempvsdepth)

NvsP <-
  ggplot(PeterPaul.chem.nutrients, aes(x = tp_ug, y = tn_ug, color = depth)) +
  geom_point() #+
  #geom_smooth(method = lm) +
  #geom_abline(aes(slope = 16, intercept = 0), lty=2) #这条线可以看出多数点是大于16:1的比例
print(NvsP)

# Exercise: Plot relationships between air quality measurements

# 1. 
# Plot AQI values for ozone (y) by PM2.5 (x), colored by site. 
# Add a line of best fit for the linear regression of these variables.
AQIrelation <- ggplot(EPAair, aes(x=PM2.5, y=Ozone, color= Site.Name))+
  geom_point()+
geom_smooth(method = lm)
  
```

### Plotting continuous vs. categorical variables

```{r, fig.height = 3, fig.width = 4, warning = FALSE}
 # Barplot + error bars
PeterPaul.nutrient.summaries <- PeterPaul.nutrients.gathered %>%
  group_by(lakename, nutrient) %>% #group by function
  summarise(sd = sd(concentration), 
            mean = mean(concentration))

Nutrientplot <- 
  ggplot(PeterPaul.nutrients.gathered) +
  geom_bar(aes(x = lakename, y = concentration, fill = as.factor(nutrient)), # why did we use fill?
           position = "dodge",  stat = "summary", fun.y = "mean")             # what's happening here? if not dodge, they will all stack in one stack.
print(Nutrientplot)

Nutrientplot2 <- 
  ggplot(PeterPaul.nutrient.summaries, aes(x = lakename, y = mean, fill = as.factor(nutrient))) + #
  geom_bar(stat = "identity", position = "dodge") + # what does the stat command do? In here we already have a number for our mean, we dont need summary here, instead we put identity.
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), # how do we specify error bars? 
                 position = "dodge")
print(Nutrientplot2)

# Are there more effective ways to produce summary stats for categories?

# Box and whiskers plot
Nutrientplot3 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_boxplot(aes(color = nutrient)) # Why didn't we use "fill"?
print(Nutrientplot3)

# Dot plot
Nutrientplot4 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_dotplot(aes(color = nutrient), binaxis = "y", binwidth = 1, 
               stackdir = "center", position = "dodge") #
print(Nutrientplot4)

# Violin plot
Nutrientplot5 <-
  ggplot(PeterPaul.nutrients.gathered, aes(x = lakename, y = concentration)) +
  geom_violin(aes(color = nutrient)) #
print(Nutrientplot5)

# Frequency polygons
# Using a tidy dataset
Nutrientplot6 <-
  ggplot(PeterPaul.chem.nutrients) +
  geom_freqpoly(aes(x = tn_ug), color = "black") +
  geom_freqpoly(aes(x = tp_ug), color = "darkblue") +
  geom_freqpoly(aes(x = nh34), color = "darkgray") +
  geom_freqpoly(aes(x = no23), color = "gray") +
  geom_freqpoly(aes(x = po4), color = "blue") 
print(Nutrientplot6)

# Using a gathered dataset
Nutrientplot7 <-   
  ggplot(PeterPaul.nutrients.gathered) +
  geom_freqpoly(aes(x = concentration, color = nutrient))
print(Nutrientplot7)

# Exercise: Plot distributions of AQI values for EPAair

# 1. 
# Create a bar chart plus standard deviation error bars for PM2.5, divided by year. 
# Create separate bars for each site. 

EPAairsummary <- EPAair %>%
  group_by(Site.Name, year) %>% 
  summarise(sd = (sd(PM2.5, na.rm=T)), 
            mean = (mean(PM2.5, na.rm = T))) 
  
PM2.5error <- 
  ggplot(EPAairsummary, aes(x = as.factor(year), y = mean, fill = as.factor(Site.Name))) + #
  geom_bar(stat = "identity", position = "dodge") +
  geom_errorbar(aes(ymin = mean-sd, ymax = mean+sd), 
                 position = "dodge")

# 2. 
# Create a new plot that better depicts the distribution of PM2.5 concentrations. 
# Divide your graph by year and site.


```
